+++ packages/polar/src/general/Origin.tsx
@@ -2,23 +2,17 @@ import { m } from 'framer-motion'
 import { useScales, ContinuousAxisScale, getClassName } from '@chsk/core'
 import { OriginProps } from './types'
 
-export const Origin = ({
-    variant = 'default',
-    setRole,
-    className,
-    style,
-    children,
-}: OriginProps) => {
+export const Origin = ({ setRole = true, className, style, children }: OriginProps) => {
     const { scales } = useScales()
     const xScale = scales.x as ContinuousAxisScale
     const yScale = scales.y as ContinuousAxisScale
     const config = { x: xScale(0), y: yScale(0), originX: '0px', originY: '0px' }
-    const compositeClassName = getClassName(variant, className)
+    const compositeClassName = getClassName('origin', className)
     return (
         <m.g
             initial={config}
             animate={config}
-            role={setRole ? variant : undefined}
+            role={setRole ? 'origin' : undefined}
             style={style}
             className={compositeClassName}
         >