+++ packages/polar/src/pie/Slice.tsx
@@ -22,7 +22,6 @@ const getSliceD = ({
 }
 
 export const Slice = ({
-    variant = 'slice',
     innerRadius,
     outerRadius,
     r = 0,
@@ -50,7 +49,7 @@ export const Slice = ({
     const [working, setWorking] = useState(false)
 
     if (d.get() !== getSliceD(sliceProps) && !working) {
-        const animateConfig = theme.Motion[variant] ?? theme.Motion.default
+        const animateConfig = theme.Motion.default
         const interpolator = interpolateObject(values, sliceProps)
         animate(0, 1, {
             ...animateConfig,
@@ -67,11 +66,11 @@ export const Slice = ({
         })
     }
 
-    const compositeClassName = getClassName(variant, className)
+    const compositeClassName = getClassName('slice', className)
     return (
         <m.path
             d={d}
-            role={setRole && variant !== 'default' ? variant : undefined}
+            role={setRole ? 'slice' : undefined}
             className={compositeClassName}
             {...props}
         />